Provides a RepeatSection Dialog component that does NOT open a dialog to edit the repeated section, it just displays everything inline.  Removing the old code which has a number of issues.
