+++ projects/sartography-workflow-lib/package.json
@@ -1,5 +1,5 @@
 {
   "name": "sartography-workflow-lib",
-  "version": "0.0.620",
+  "version": "0.0.621",
   "dependencies": {}
 }
