--- projects/sartography-workflow-lib/src/lib/modules/forms/repeat-section-dialog/repeat-section-dialog.component.html
@@ -1,36 +1,0 @@
-<div
-  class="mat-typography"
->
-  <div mat-dialog-content>
-    <div fxLayout="row" fxLayout.lt-md="column">
-      <h1>{{ data.title }}</h1>
-      <span fxFlex></span>
-      <button
-        mat-button
-        id="highlight-required-fields-in-dialog"
-        (click)="highlightRequiredFields(data.fields)"
-        color="warn"
-      >*Required</button>
-    </div>
-    <formly-form
-      [fields]="data.fields"
-      [model]="data.model"
-      (change)="updateDisableSave()"
-      [options]="data.options"
-    ></formly-form>
-  </div>
-  <div mat-dialog-actions>
-    <button
-      mat-flat-button
-      color="primary"
-      (click)="onSubmit()"
-      [ngClass]="{'repeat-section-dialog-save': true}"
-      [disabled]="!noErrors()"
-    >Save</button>
-    <button
-      mat-button
-      (click)="onNoClick()"
-      [ngClass]="{'repeat-section-dialog-cancel': true}"
-    >Cancel</button>
-  </div>
-</div>
