--- projects/sartography-workflow-lib/src/lib/modules/forms/repeat-section-dialog/repeat-section-dialog.component.scss
@@ -1,3 +1,0 @@
-#highlight-required-fields-in-dialog {
-  text-align: right;
-}
