--- projects/sartography-workflow-lib/src/lib/modules/forms/repeat-section/repeat-section.component.html
@@ -1,51 +1,0 @@
-<div
-  *ngIf="field && field.fieldArray"
-  [ngClass]="{'container': true, 'invalid': !field.formControl.valid}"
-  fxLayout="column"
-  fxLayoutGap="1em"
->
-  <mat-error *ngIf="!form.valid">
-  <formly-validation-message [field]="field"></formly-validation-message>
-  </mat-error>
-
-  <mat-card *ngFor="let fg of field.fieldGroup; let i = index;" class="repeat">
-    <mat-card-content fxLayout="row">
-      <div fxFlex="70">
-        <lib-form-printout *ngFor="let f of fg.fieldGroup" [field]="f" [model]="f.model"></lib-form-printout>
-      </div>
-      <div
-        fxFlex="30"
-        fxLayout="row"
-        fxLayoutAlign="end start"
-      >
-        <button
-          (click)="openDialog(i, fg)"
-          mat-icon-button
-          type="button"
-        >
-          <mat-icon>edit</mat-icon>
-        </button>
-        <button
-          *ngIf="!field.templateOptions.editOnly"
-          (click)="confirmDelete(i)"
-          mat-icon-button
-          type="button"
-        >
-          <mat-icon>close</mat-icon>
-        </button>
-      </div>
-    </mat-card-content>
-  </mat-card>
-</div>
-<div *ngIf="field && field.templateOptions" class="repeat-action">
-  <button
-    *ngIf="!field.templateOptions.editOnly"
-    (click)="openDialog(field.fieldGroup ? field.fieldGroup.length : 0)"
-    color="accent"
-    mat-flat-button
-    type="button"
-  >
-    <mat-icon>add</mat-icon>
-    Add {{ field.templateOptions.buttonLabel }}
-  </button>
-</div>
