+++ projects/sartography-workflow-lib/src/public-api.ts
@@ -19,8 +19,7 @@ export * from './lib/modules/forms/help-wrapper/help-wrapper.component';
 export * from './lib/modules/forms/markdown-description-wrapper/markdown-description-wrapper.component';
 export * from './lib/modules/forms/panel-wrapper/panel-wrapper.component';
 export * from './lib/modules/forms/radio-data-field/radio-data-field.component';
-export * from './lib/modules/forms/repeat-section-dialog/repeat-section-dialog.component';
-export * from './lib/modules/forms/repeat-section/repeat-section.component';
+export * from './lib/modules/forms/repeat-type/repeat-type.component';
 export * from './lib/modules/forms/validators/email.regex';
 export * from './lib/modules/forms/validators/email.validator';
 export * from './lib/modules/forms/validators/formly.validator';
